// frontend/src/hooks/useAddPlans.d.ts
import { AxiosError } from "axios";
interface PlanFormData {
    _id?: string;
    name: string;
    type: "paid" | "free";
    description: string;
    regularAmount: number;
    discountAmount: number;
    benefits: string[];
    isListed: boolean;
    duration: '1 month' | '3 months' | '6 months' | '9 months' | '12 months';
    createdAt?: Date;
    isDeleted?: boolean;
    roomBenefit: number;
}
export declare const useAddPlan: () => {
    data: undefined;
    error: null;
    isError: false;
    isIdle: true;
    isLoading: false;
    isSuccess: false;
    status: "idle";
    mutate: import("@tanstack/react-query").UseMutateFunction<unknown, AxiosError<unknown, any>, PlanFormData, unknown>;
    reset: () => void;
    context: unknown;
    failureCount: number;
    failureReason: AxiosError<unknown, any> | null;
    isPaused: boolean;
    variables: PlanFormData | undefined;
    mutateAsync: import("@tanstack/react-query").UseMutateAsyncFunction<unknown, AxiosError<unknown, any>, PlanFormData, unknown>;
} | {
    data: undefined;
    error: null;
    isError: false;
    isIdle: false;
    isLoading: true;
    isSuccess: false;
    status: "loading";
    mutate: import("@tanstack/react-query").UseMutateFunction<unknown, AxiosError<unknown, any>, PlanFormData, unknown>;
    reset: () => void;
    context: unknown;
    failureCount: number;
    failureReason: AxiosError<unknown, any> | null;
    isPaused: boolean;
    variables: PlanFormData | undefined;
    mutateAsync: import("@tanstack/react-query").UseMutateAsyncFunction<unknown, AxiosError<unknown, any>, PlanFormData, unknown>;
} | {
    data: undefined;
    error: AxiosError<unknown, any>;
    isError: true;
    isIdle: false;
    isLoading: false;
    isSuccess: false;
    status: "error";
    mutate: import("@tanstack/react-query").UseMutateFunction<unknown, AxiosError<unknown, any>, PlanFormData, unknown>;
    reset: () => void;
    context: unknown;
    failureCount: number;
    failureReason: AxiosError<unknown, any> | null;
    isPaused: boolean;
    variables: PlanFormData | undefined;
    mutateAsync: import("@tanstack/react-query").UseMutateAsyncFunction<unknown, AxiosError<unknown, any>, PlanFormData, unknown>;
} | {
    data: unknown;
    error: null;
    isError: false;
    isIdle: false;
    isLoading: false;
    isSuccess: true;
    status: "success";
    mutate: import("@tanstack/react-query").UseMutateFunction<unknown, AxiosError<unknown, any>, PlanFormData, unknown>;
    reset: () => void;
    context: unknown;
    failureCount: number;
    failureReason: AxiosError<unknown, any> | null;
    isPaused: boolean;
    variables: PlanFormData | undefined;
    mutateAsync: import("@tanstack/react-query").UseMutateAsyncFunction<unknown, AxiosError<unknown, any>, PlanFormData, unknown>;
};
export {};